[2-4] Moved auth.service file to service directory, removed comments
