+++ exalearn-ui/src/app/guards/auth.guard.ts
@@ -1,7 +1,7 @@
-import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
+import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
 import { Injectable } from '@angular/core';
-import { AuthService } from '../auth.service';
+import { AuthService } from '../services/auth.service';
 
 @Injectable({ providedIn: 'root' })
 export class AuthGuard implements CanActivate {
@@ -12,13 +12,6 @@ export class AuthGuard implements CanActivate {
 			if (isAuth) {
 				this.router.navigateByUrl('/main');
 			}
-			// else {
-			//   this.router.navigate(['/'], {
-			//     queryParams: {
-			//       auth: false
-			//     }
-			//   })
-			// }
 		});
 		return true;
 	}
